requirements.txt
anyio==3.6.2
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.9.24
contourpy==1.0.5
cycler==0.11.0
emoji==2.1.0
fonttools==4.37.4
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
isOdd==0.1.2
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
progress==1.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.0a4
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
roman==3.3
six==1.16.0
sniffio==1.3.0
tornado==6.2
tzdata==2022.5
tzlocal==4.2
